Work around GCC bug to support GCC 5.0-8.0
GCC in these versions did not like unconditional throw statements in
constexpr function, even if they weren't evaluated. I'm not too sure
about the background of this bug, but it was ultimately deemed a bug and
fixed in GCC 9.0. For sake of compatability though, the code was
refactored to support older GCC that contained this bug.